+++ protected/models/interfaces/IAntrag.php
@@ -12,6 +12,10 @@ abstract class IAntrag extends GxActiveRecord
 	public static $STATUS_ABGELEHNT = 5;
 	public static $STATUS_MODIFIZIERT_ANGENOMMEN = 6;
 	public static $STATUS_MODIFIZIERT = 7;
+	public static $STATUS_UEBERNAHME = 8;
+	public static $STATUS_ERLEDIGT = 9;
+	public static $STATUS_UEBERWEISUNG = 10;
+	public static $STATUS_ABSTIMMUNG = 11;
 	public static $STATI = array(
 		-2 => "Gelöscht",
 		-1 => "Zurückgezogen",
@@ -21,11 +25,14 @@ abstract class IAntrag extends GxActiveRecord
 		3  => "Eingereicht",
 		4  => "Angenommen",
 		5  => "Abgelehnt",
-		6  => "Modifiziert angenommen",
-		7 => "Modifziert",
+		6  => "Modifizierte Übernahme",
+		7  => "Modifziert",
+		8  => "Übernahme",
+		9  => "Erledigt",
+		10 => "Überweisung",
+		11 => "Abstimmung",
 	);
 	public static $STATI_UNSICHTBAR = array(0, 2, -2);
-
 
 
 	/**